// strimzi/tls-clients/kafkajs-client/src/utilis/config/interface.ts
export interface Config {
  LOG_LEVEL: string

  KAFKA_BROKERS: string

  KAFKA_CLIENT_CONNECTION_TIMEOUT: number
  KAFKA_CLIENT_REQUEST_TIMEOUT: number
  KAFKA_CLIENT_RETRIES: number
  KAFKA_CLIENT_LOG_LEVEL: number

  KAFKA_CONSUMER_SESSION_TIMEOUT: number
  KAFKA_CONSUMER_HEARTBEAT_INTERVAL: number
  KAFKA_CONSUMER_RETRIES: number
  KAFKA_CONSUMER_TOPIC: string
  KAFKA_CONSUMER_GROUP_ID: string
  KAFKA_PRODUCER_TOPIC: string

  MESSAGING_CLIENT_DOC: string

  KAFKA_SSL_CONFIG_ACTIVE: boolean

  KAFKA_CLIENT_CERT_PATH: string
  KAFKA_CLIENT_KEY_PATH: string
}